Refine the change to ignore .docc files rather than emit warnings to avoid having multiple definitions for the same file extension
Ignore .docc files rather than emitting warnings, since it's useful to have them in packages without warnings.

Unlike other Xcode-specific file types, such as Storyboards and Asset Catalogs, the .docc bundles are not needed for correctness during the build, and should therefore not trigger warnings.

This is a refinement of the fix in #3609.

rdar://78133445