[Collections] Don't queue populateTargetTrie if there is no data (#3634)
Motivation:
`populateTargetTrie`, potentially because it's done in a background queue, causes a lot of random crashes in tests (not just in SwiftPM but other projects as well). The exception happens inside SQLite C library and it's not yet clear what the real cause is. However, since most tests do not use the package collections feature at all, `populateTargetTrie` is basically a no-op and therefore doesn't need to be queued.

Modification:
Queue `populateTargetTrie` only if there is package collection data.

This code change doesn't actually fix anything but may perhaps reduce the chance of crashing.

rdar://80840989