PackageCollectionsSigning: correct `BoringSSLCertificate.keyType(of:)` (#3679)
This function takes a named argument which is the `EVP_PKEY` and checks
the underlying public key type (expected to be one of RSA or EC).
However, instead of checking the key type of the given public key, it
was always returning the key type of the underlying certificate.  This
happened to give the correct result as the current usage always used the
same object to get the key.  This corrects that and uses the proper
accessors for the value.