Allow skipping overrides for `swiftc`
If we are cross-compiling, the compiler being passed to `bootstrap` will be assumed to be for the target, not the host, so we should skip passing it to the first stage build and fall back to the installed toolchain.

In the future, it might be interesting to have separate `swiftc` overrides for host and target platform.