Implement proposed adjustments to SE-0303 to use `@main` for plugins
This is an implementation of the proposed adjustments to SE-0303 in https://github.com/apple/swift-evolution/pull/1434.

This changes the entry point to use `@main` rather than top-level code, and introduces protocols for the entry points corresponding to the various plugin capabilities.  It gets rid of the globals in favour of plugin parameters and return values (except for diagnostics, which are still collected using a process-wide list).