refactor diagnostics
motivation: refactor the diagnostics frameworks to seperate between emitting diagnostics and collecting/handling them, this helps creating an API to allow emitting diagnostics from any point at the code without having to pass in a DiagnosticsEngine everywhere

changes:
* add newe ObservabilitySystem API to abstract features like diagnostics, metrics and tracing
* add new DiagnosticsEmitter API design to [only] emit diagnostics
* adjust callsites to always create DiagnosticsEngine using the ObservabilitySystem so that the system are bridged
* convert GraphLoading subsystem to use DiagnosticsEmitter instead of DiagnosticsEngine
* move the Diagnostic printing code closer the Diganostic handling code in SwiftTool and clean it up
* adjust all tests and test utilities to use ObservabilitySystem

note: the code includes several API to allow bridging between the DiagnosticsEngine using the ObservabilitySystem APIs so that we can transition in steps and control the scope of the cahnges. The APIs are marked as deprecated to help trasnition, but are fairly noisy until we make more progress.