libSwiftPM clients should be able to choose which PackageDescription library path to use based on the package
This allows a choice not only based on the tools version but also based on the package path.  The client is then free to substitute specially built versions of PackageDescription on a package-by-package basis, if needed.  We will be able to use this to remove some of the special-casing from the general libSwiftPM logic.

rdar://82422039