introduce support for local archived binary targets.
To significantly speed up builds and make them more reliable (especially in a CI environment) it makes sense to precompile and put 3rd party dependencies under source control to later consume them as local binary targets.
To facilitate managing and maintaining those binary targets it would make sense to consume them in an archived form because that would decrease the number of files and the general repository size a lot (sometimes in orders of magnitude), not speaking of improving diff commits when updating versions of the targets.
That is why the main motivation of this PR is to introduce archive support for local binary targets to address the issues mentioned above.