Fix a unit test that is writing to the checkout of SwiftPM itself.
This was never a good thing to do, but now that we have v2 of Package.resolved, it is an actively harmful thing to do if the IDE or toolchain being used to build SwiftPM and run the tests is older (e.g. using the 5.5 toolchain but working on main).

There's not need at all to run this test against SwiftPM itself — a sample package created for the purpose in the usual temporary directory is all that's needed.

rdar://83222890