Extend the plugin API by passing down more information about the package graph
The original TargetBuildContext input structure was focused on build tools, which was the subject of SE-0303.  In order to prepare for new kinds of plugins such as custom commands, which have much more varied purposes and need a more diverse set of input information, we generalize this structure into a PluginContext that contains information about a whole package graph.

The graph — which is a directed acyclic graph and not trivially encodable into a transport format — is flattened and serialized into the plugin input JSON, and is reconstituted on the plugin side.  Path deduplication is used to reduce the amount of information sent over the wire.

A compatibility layer lets old plugins keep working with the old TargetBuildContext structure.  This is preliminary pending approval of an upcoming pitch and evolution proposal.