`ProductFilter` type is not `Decodable` for the `everything` case (#3775)
`ProductFilter.everything` is encoded as the `.none` case of a `[String]?`, which gets encoded by the SingleValueEncodingContainer as JSON `null`.  But then `SingleValueDecodingContainer` doesn't decode the `null` as the `.none` case of `[String]?`, instead throwing an error.  This change uses `decodeNil` to check for `null`, but it seems to me that there might be a bug in SingleValue{Encoding,Decoding}Container here, since it doesn't seem symmetrical.  I would expect consistent handling of `[String]?` for encoding and decoding.

rdar://83647022