Disuse unversioned triples on non-Darwin platforms. (#3576)
The target flag should take the versioned triple, otherwise on platforms
with versioned triples, the standard library won't be found.

On Darwin, the unversioned triple should still be used throughout. This
necessitates special-casing in the bootstrap script and when making
subdirectories during the build.

This platform-specific branch is encapsulated in a small extension
on Triple in swiftpm. All tests that use the `tripleString` to construct
the `.build` subdirectory are updated accordingly.

See TSC pr #229.