rdar://83703185 (`swift build` isn't passing `-fmodules` to C files in targets that contain one or more C++ files)
SwiftPM's handling of mixed C and C++ targets has several issues.  One is fallout from the fixing of rdar://83416799, which caused Clang targets that contain even a single C++ file to be built without Clang modules enabled.  This changes it to be on a per-Clang-invocation basis.  To be technically correct it should look at any `-x` argument and not just the filename suffix, but that is the subject of a future PR.

To avoid revlock with clients such as SourceKit-LSP, this PR preserves the old behaviour if the caller of `BuildPlan.basicArguments()` doesn't specify whether to use C++ flags.  This will be unwound once clients have switched over to always passing this option.