Implement SE-0325 Additional Package Plugin APIs (#3758)
* Extend the plugin API by passing down more information about the package graph

The original TargetBuildContext input structure was focused on build tools, which was the subject of SE-0303.  In order to prepare for new kinds of plugins such as custom commands, which have much more varied purposes and need a more diverse set of input information, we generalize this structure into a PluginContext that contains information about a whole package graph.

The graph — which is a directed acyclic graph and not trivially encodable into a transport format — is flattened and serialized into the plugin input JSON, and is reconstituted on the plugin side.  Path deduplication is used to reduce the amount of information sent over the wire.

A compatibility layer lets old plugins keep working with the old TargetBuildContext structure.  This is preliminary pending approval of an upcoming pitch and evolution proposal.

* Change Product and Target to protocols, and change the classes to structs.  Since structs cannot be nested in protocols, `Package.Dependency` moves out to a top-level `PackageDependency`, and similarly `Target.Dependency` moves out to `TargetDependency`.  Also define identifiers for Package, Product, and Target since they are now value types.  We cannot make them conform to `Identifiable` and still use arrays of `Product` and `Target` due to the Self requirement of `Identifiable`.  We can maybe resolve this in the future.  For now we use the ID that we already have from the model that we were handed from libSwiftPM as the identifier, but the exact value doesn't matter.

* Move the `builtProductsDirectory` in the package plugin context to be a private property.

The original intent was that it could be used in forming paths to built executables, but this is instead done through the `tools` to paths mapping.

* Bring draft implementation in line with proposal:

- add Origin to Package struct
- add SourceModuleTarget protocol, and separate SwiftSourceModuleTarget and ClangSourceModuleTarget
- fill in BinaryArtifactTarget and SystemLibraryTarget
- add ToolsVersion to Package struct

* Add BuildEnvironment to plugin evaluation.

* Adjust to changes to binary targets to have a `unknown` artifact type

* Add support for `pkgConfig` for system library targets.