Make the communication between the plugin host and the plugin slightly more sophisticated, in preparation for being able to send messages back and forth.
This approach uses `stdout` for JSON information from the plugin and `stderr` for freeform text from the plugin rather than a specialized zero byte denoting the start of the output JSON.

We also close `stdin` to avoid having the plugin block forever in case it tries to read from `stdin` (console input to the plugin isn't available).

No change in functionality except that trying to read from `stdin` from the plugin no longer blocks it.