Fix archive-source command
Motivation:
Source archives created using the `archive-source` command doesn't
include top-level directory.

Modification:
The trailing slash `/` in the `--prefix` argument to `git archive`,
which is the underlying command for `archive-source`, is required to
create top level directory. Without it the prefixed to all the files and
dirs in the archive.