Package fingerprint storage
Motivation:
Provide trust-on-first-use (TOFU) security model for package dependencies. SwiftPM will record locally the "fingerprint" of a package version when it is first downloaded and ensure the fingerprint remain the same in subsequent downloads.

For package downloaded from registry, the fingerprint is the source archive checksum. For source control, it is the git revision.

This PR only adds storage APIs for writing and reading package fingerprints and an implementation that uses local file system. The integration of fingerprint storage into the download workflow(s), i.e. TOFU implementation, will come in a separate PR.

Modifications:
- Add `PackageFingerprint` module.
- Add `PackageFingerprintStorage` protocol that defines APIs for reading and writing fingerprints.
- Add `FilePackageFingerprintStorage` which is an implementation based on file system.