Initial implementation of Command Plugins (#3855)
Initial implementation of support for SwiftPM Command Plugins.

https://forums.swift.org/t/pitch-package-manager-command-plugins/53172

This feature is experimental until approved, so declaring command plugins requires a tools version of `999.0` in the package manifest, and invoking them requires `SWIFTPM_ENABLE_COMMAND_PLUGINS` to be set to `1` in the environment of `swift` `package`.

This implementation temporarily uses a `swift package plugin <verb>` form for invoking command plugins. Per the pitch and proposal, plugin-defined commands are intended to be invocable using the shortcut `swift package <verb>` when the feature is complete (there is discussion of even supporting `swift <verb>` but that is not part of the current proposal).

This feature is being developed incrementally across this commit and follow-on commits. This commit includes all the proposed changes to `PackageDescription` and `PackagePlugin`. Remaining items of the full feature will then be landed in future PRs. The `SWIFTPM_ENABLE_COMMAND_PLUGINS` guard flag will only be removed once the evolution proposal has been approved and the feature has been fully implemented including any amendments.

Specific modifications in this commit:

- add the new enum cases for the command plugin capability to PackageDescription
- add serialization of those cases in PackageDescriptionSerializer
- deserialize them in the ManifestJSONLoader and add to the package model
- add manifest source generation for the command capability
- add ability to describe packages that have command plugins
- add a unit test to check loading a manifest with a command capability
- add protocol declarations for command plugin entry point in the plugin API
- ability to invoke the command plugin in libSwiftPM and unit tests
- add a unit test to check round trip calling into the plugin
- add ability to invoke the command plugin from CLI
- made the host-to-plugin communication be bidirectional and interactive
- ability for plugin to call back into SwiftPM for services
- ability to create symbol graphs for a target when a plugin requests it
- ability to run a build when the plugin requests it

Remaining to implement:

- request approval to run command plugins that need special permissions
- implement the option to list the available command plugins
- support qualifiers for ambiguous plugin commands
- ability to run a test when the plugin requests it
- build any executables needed by command plugins before invoking the plugin
- provide access to toolchain executables in the map sent to plugins
- add async to all the plugin APIs (in separate PR, depends on back deployment)

Some of these are not specific to command plugins but rather apply to all plugins in general.