Make sure that NSTemporaryDirectory() is included when `writableTemporaryDirectory` is set.
It seems that `TSCBasic.determineTempDirectory()` lets environment variables override the temporary directory, which might be alright for the process itself but not necessarily for spawned subprocesses.  If the subprocess uses NSTemporaryDirectory(), then that still won't be covered by the sandbox.

This change makes sure that it is covered by the sandbox, and uses a set to avoid duplicates (which wouldn't actually hurt but is inelegant).

rdar://86150592