Optionally use a per-package filesystem for `PackageBuilder`
Currently, we have the ability to provide a custom file system to the whole workspace and we partially use git-backed file system during package resolution. This change aims to allow using a custom file system when running `PackageBuilder` as well which e.g. allows to load a graph without having full checkouts of every package.

To utilize this feature, clients of `Workspace` can pass a custom `PackageContainerProvider` and use the new `CustomPackageContainer` protocol.