Test creating 'fingerprints' dir
Motivation:
Source compat test continues to fail even with https://github.com/apple/swift-package-manager/pull/3928:

```
"You don’t have permission to save the file “fingerprints” in the folder “security”
```

https://ci.swift.org/job/swift-PR-source-compat-suite/5709/artifact/

Not sure how/why, but it seems like `security` directory is created just fine but not the `fingerprints` sub directory (`fingerprints` is a directory not a file, and the code add a file to the `fingerprints` dir for each downloaded package).

Modifications:
Test creating `fingerprints` directory instead of `security`, and if that fails, disable TOFU feature.