Errors thrown at top level of plugin should be reported as diagnostics
When a plugin throws an uncaught error, it should be reported as a diagnostic rather than causing the standard Swift output.

As part of making this work well, we also simplify the way the result code is received from the plugin.  It was a bit awkward to have two result codes (the one from the last message as well as the exit code from the process) so now we just use the exit code.  This is also most robust to any plugin that decides to do "exit(1)" on its own from inside the code.  If/when we have long-running multi-action plugins we can revisit this.

We also make sure that at least one error message is emitted from any plugin that exits with an error code but that does not emit an error diagnostic.

This also reworks the unit test for command invocations to be more scalable.