Fix an ordering issue of a print statement and a diagnostic arriving right at the exit of a plugin
There's no strong ordering between the free form output of a plugin and the diagnostics it emits, but for things like thrown exceptions etc, it makes a bit more sense to show the diagnostic last.

Perhaps the test shouldn't really depend on the order here — we will see if this addresses the issue in the CI (it seems specific to Linux and I was never able to reproduce it in Docker).

I also made the unit test text a little less whimsical, and got rid of the emoji.