Support `--allow-writing-to-package-directory` as described in SE-0332 and `--allow-writing-to-directory` to allow arbitrary directories (#3972)
This adds `--allow-writing-to-package-directory` as described in SE-0332.  It also adds `--allow-writing-to-directory`, which takes a relative or absolute path as an argument, and also adds `--block-writing-to-temporary-directory` to provide control over whether the temporary directories can be written to.

It also adds in the error to message that is shown if writing to the package directory isn't allowed, which includes the stated reason the plugin provided for why it needs this access.

rdar://86786404
(cherry picked from commit a0c0d1fcc4886894947fc16d2b42e43b5c3c97e9)