Rework the SymbolGraphExtract so it's more suitable for being called in the context of plugins.
This includes parameterizing the output that's printed and the output path, and applying the extractor to a target at a time and not always to all the targets in a package.  The original behavior for the `dump-symbol-graph` command is unmodified, but some of the specifics are hoisted into the DumpSymbolGraph command implementation rather than in the reusable SymbolGraphExtract type.