Make the symbol graph extraction requested by plugins more robust (#3993)
* Rework the SymbolGraphExtract so it's more suitable for being called in the context of plugins.

This includes parameterizing the output format and the output path, as well as the diagnostics that are printed, and applying the extractor to a target at a time and not always to all the targets in a package.  The original behavior for the `dump-symbol-graph` command is unmodified, but some of the specifics are hoisted into the DumpSymbolGraph command implementation rather than in the reusable SymbolGraphExtract type.

* Make the symbol graph generation more robust by using the refactored SymbolGraphExtract:

- only generate symbol graph information for the specified target
- use a unique symbol graph output directory for each package/target combination
- added a unit test to check these things.

* Skip the test in a bootstrap situation where we're using a Swift compiler that hasn't built the `swift-symbolgraph-extract` tool.  There is a `SWIFT_SYMBOLGRAPH_EXTRACT` env var that could be used to point it another binary, but on platforms other than macOS we don't have an existing Xcode or anything that would be guaranteed to contain this tool.  This test should still run on full builds and on self tests, where the tool is available.

rdar://87234110&86785765
(cherry picked from commit 850c772727825a9d09a7febfd6d3819ab52c1fe8)