Enforce read-only-ness of package directory even when it's inside writable directories (#3996)
* Enforce read-only-ness of package directory even when it's inside /tmp and other writable directories

The sandbox supports a list of `writableDirectories`, which are treated as path component prefixes after package resolution.  The are applied to the sandbox rules after the default-deny and everything-is-read-only rules, carving out cones of writability in specific areas of the file system.

But if the package directory is under one of these directories that is allowed by default, then it is implicitly writable even in the absence `--allow-writing-to-package-directory`.  It also makes unit tests awkward, since they usually create temporary test fixtures inside the temporary directory.

This adds a set of `readOnlyDirectories` in addition to the writable ones, so that the package directory remains unwritable regardless of where it is located.

It also adds some documentation in the Sandbox source file, and it removes a somewhat confusing `--block-writing-to-package-directory` flag that was added mostly for unit tests but that is no longer needed.

(cherry picked from commit c117c86d248713a0ffc777bc190d16f0373f54f8)