Allow running tests from a command plugin (#4002)
This wraps up the last piece for SE-0332, including support for filtering and for capturing code coverage while the tests are running.

This uses most of the same implementation that existed in a very specify way in SwiftTestTool.swift, factoring it out into a new shared TestingSupport.swift file and making it more generic. Future refactoring should move it down to a new module that provides the shared test support for use from everywhere.