Package command plugins should have the same initial working directory as the invocation
Motivation:  This allows relative paths passed as parameters to work.  For build tool plugins we pass in the package directory so that relative paths in configuration files work there as well.  This also matches the behavior of many build systems.  The proposal didn't specify the initial working directory and usually all paths are absolute, but having predictable initial working directories follows the principle of least surprise.

rdar://86831942