[5.6] Don't emit a "this package dependency is unused" warning for packages that contain at least one command plugin product (#4028)
An existing warning about dependencies on packages from which no products are used shouldn't apply if the package has any command plugin products.  The warning assumes that the only use of products is to link against them or use them as build tool plugins, but that's not correct for command plugins, where the dependency is only used to make the plugin available for the user to invoke.

rdar://86787186
(cherry picked from commit cd9dc448385baf992f4d7190e2654ef6ce407c6b)