Emit output and diagnostics from build tool plugin invocations as diagnostics
Diagnostics are emitted as warnings or errors depending on what the plugin specified, and general print statements from the plugin are emitted as the `.info` level of diagnostics.  A future improvement could improve this by adding a `.output` level between `.info` and `.warning` that doesn't require `-v`.

This PR also adds some other information to the build result structure that is useful to clients of libSwiftPM.