[5.6] Make the `@main` entry points for plugins be `async` so they can call other `async` APIs (#3842)
This also adds a `-rpath` flag when compiling the plugin to run on macOS earlier than 12.0, so that the back deployment libraries in the compiler's toolchain can be found.

Also add a test helper function to determine whether the current environment actually supports concurrency (which isn't straightforward to answer by inspection in the case of back deployment on macOS), and add uses of this function to conditionally skip all the plugin tests that require async await.  This allows SwiftPM to still be built with older SwiftPM versions.

rdar://83206272

(cherry picked from commit e78bc7c717911e805f3416418c7c39dbc39a2f09)