Pass along `--target` and `--product` flags from `swift package` invocation to command plugin's `arguments` array
As part of this we simplify the command plugin method to not have a `targets` parameter but to instead get all the inputs in `arguments`.

Any `--target` or `--product` flags from the `swift package` invocation are now prepended as `--target` and `--product` parameters in the plugin arguments, respectively, and the ArgumentExtractor helper type makes it easier to get at this information from the plugin.  Also add methods to make it easier to look up products and targets in the plugin context's package.

rdar://86786279