Fix issues caused by having the `plugin` subcommand be the default subcommand
There were a few undesirable issues caused by a combination of a `swift-argument-parser` issue and making the `plugin` subcommand be the default subcommand so that the `plugin` subcommand could be omitted:
 - when you invoke `swift package` without any arguments it starts resolving the package graph
 - when you make a typo like `swift package resove` it emits an error that it did not find any plugin (`error: No command plugins found for ‘resove’`)
 - the `--list` option of the `plugin` subcommand incorrectly applied even when `plugin` was omitted
 - `swift package plugin --help` yields no help
 - `swift package plugin` yields `error: No plugin command name specified`

This change introduces a new separate default subcommand that has behavior suitable for omitting the `plugin` subcommand:
 - if there are no parameters, it prints the help
 - if there is an unknown parameter, it prints an error message that takes into account that it might not be a command
 - it doesn't define options like `--list` that only make sense in the context of a `plugin` subcommand

This still leaves the problem of having options defined by `swift package`, such as `--verbose`, getting "stolen" even if they are on the right hand side of the plugin name.  Fixing that depends on a fix in `swift-argument-parser` that is planned to be included in an upcoming bugfix tag, but this is still incremental progress.